some cleanup and improvements
