+++ src/buffer_allocator.rs
@@ -1,79 +1,111 @@
-use crate::port_type::PortType;
-use std::cmp::PartialEq;
-use std::marker::PhantomData;
+///! Buffer allocation internals v2.
+///!
+///! The [BufferAllocator] is implemented using a list of
+///! stacks, with one stack for each port type. When a
+///! new buffer is required, we first try and pop a buffer
+///! off the stack. If it none are available, we allocate
+///! a new buffer. When a buffer is released, it is
+///! pushed to the top of the corresponding stack.
+///!
+///! There is some additional bookkeeping required for the
+///! buffers. [BufferRef]s contain a `ref_count` field which
+///! tracks the number of edges that still need the buffer
+///! to be alive before it can be safely released. The
+///! `generation` field is kept around for visualization
+///! of the assigned buffers during debugging.
+///!
+///! Finally, the engine using the graph needs to know
+///! the maximum number of buffers for each port type
+///! to allocate during its prepare for playback operation.
+///! We track this by counting each time a new buffer
+///! is allocated for a type in the `counts` list.
+///!
+///! Since it is not valid for the buffer allocator to
+///! keep allocating after the `counts` field has been
+///! consumed, we require consuming `self` to retrieve it.
+use std::rc::Rc;
 
-/// An abstract buffer. Represents an allocated collection of data
+#[cfg(feature = "serialize")]
+use serde::{Deserialize, Serialize};
+
+/// A reference to an abstract buffer during buffer allocation.
+#[cfg_attr(feature = "serialize", derive(Serialize, Deserialize))]
 #[derive(Copy, Clone, Debug)]
-pub struct Buffer<P>
-where
-    P: PortType,
-{
-    /// The ID of the buffer
-    pub buffer_id: usize,
-    /// The type of data contained in this buffer
-    pub port_type: P,
+pub struct BufferRef {
+    /// The index of the buffer
+    pub idx: usize,
+    /// The type index of the buffer
+    pub type_idx: usize,
+    /// The generation, or the nth time this buffer has
+    /// been assigned to a different edge in the graph.
+    pub generation: usize,
 }
 
-/// The buffer allocator is a stack based allocator that reuses buffers as needed
-#[derive(Debug)]
-pub struct BufferAllocator<P> {
-    buffer_count_stacks: Vec<(usize, Vec<usize>)>,
-    _phantom_port_type: PhantomData<P>,
-}
-
-impl<P> BufferAllocator<P>
-where
-    P: PortType + PartialEq,
-{
-    /// Empty/reset the allocator. Invalidates all previously allocated buffers.
-    pub fn clear(&mut self) {
-        for (c, s) in self.buffer_count_stacks.iter_mut() {
-            *c = 0;
-            s.clear();
-        }
-    }
-
-    /// Acquire a new unique buffer.
-    pub fn acquire(&mut self, type_: P) -> Buffer<P> {
-        let type_index = type_.id();
-        let (count, stack) = &mut self.buffer_count_stacks[type_index];
-
-        if let Some(index) = stack.pop() {
-            Buffer {
-                buffer_id: index,
-                port_type: type_,
-            }
-        } else {
-            let buffer = Buffer {
-                buffer_id: *count,
-                port_type: type_,
-            };
-            *count += 1;
-            buffer
-        }
-    }
-
-    /// Release a buffer to be reused later.
-    pub fn release(&mut self, ref_: Buffer<P>) {
-        let type_index = ref_.port_type.id();
-        let stack = &mut self.buffer_count_stacks[type_index].1;
-        stack.push(ref_.buffer_id);
-    }
-}
-
-impl<P> Default for BufferAllocator<P>
-where
-    P: PortType,
-{
-    fn default() -> Self {
-        let num_types = P::NUM_TYPES;
-        let mut buffer_count_stacks = Vec::<(usize, Vec<usize>)>::new();
-        for _ in 0..num_types {
-            buffer_count_stacks.push((0, Vec::new()));
-        }
+impl BufferRef {
+    /// Create a new BufferRef
+    pub fn new(idx: usize, type_idx: usize, generation: usize) -> Self {
         Self {
-            buffer_count_stacks,
-            _phantom_port_type: PhantomData::default(),
+            idx,
+            type_idx,
+            generation,
         }
+    }
+}
+
+/// An allocator for managing and reusing [BufferRef]s.
+#[cfg_attr(feature = "serialize", derive(Serialize, Deserialize))]
+#[derive(Clone, Debug)]
+pub struct BufferAllocator {
+    /// A list of free buffers that may be reallocated, one list for
+    /// each different port type.
+    pub free_lists: Vec<Vec<FreeListEntry>>,
+    /// A list of the maximum number of buffers used for each port type.
+    pub counts: Vec<usize>,
+}
+
+/// A small helper struct for tracking the index and generation
+/// of data in the free lists
+#[cfg_attr(feature = "serialize", derive(Serialize, Deserialize))]
+#[derive(Clone, Debug)]
+pub struct FreeListEntry {
+    pub idx: usize,
+    pub generation: usize,
+}
+
+impl BufferAllocator {
+    /// Create a new allocator, `num_types` defines the number
+    /// of buffer types we may allocate.
+    pub fn new(num_types: usize) -> BufferAllocator {
+        Self {
+            free_lists: vec![vec![]; num_types],
+            counts: vec![0; num_types],
+        }
+    }
+
+    /// Acquire a new buffer with a given type index. Panics if
+    /// the type index is out of bounds.
+    pub fn acquire(&mut self, type_idx: usize) -> Rc<BufferRef> {
+        let entry = self.free_lists[type_idx].pop().unwrap_or_else(|| {
+            let idx = self.counts[type_idx] + 1;
+            self.counts[type_idx] = idx;
+            FreeListEntry { idx, generation: 0 }
+        });
+        Rc::new(BufferRef::new(entry.idx, type_idx, entry.generation))
+    }
+
+    /// Release a BufferRef.
+    pub fn release(&mut self, buffer_ref: Rc<BufferRef>) {
+        if Rc::strong_count(&buffer_ref) == 1 {
+            self.free_lists[buffer_ref.type_idx].push(FreeListEntry {
+                idx: buffer_ref.idx,
+                generation: buffer_ref.generation + 1,
+            });
+        }
+    }
+
+    /// Consume the allocator to return the maximum number of buffers used
+    /// for each type.
+    pub fn num_buffers_per_type(self) -> Vec<usize> {
+        self.counts
     }
 }
