+++ src/output_ir.rs
@@ -1,10 +1,13 @@
 //! Output data structures from the audio graph compiler.
-//!
-use crate::input_ir::Edge;
+
+use crate::input_ir::{Edge, NodeID, PortID};
+
+#[cfg(feature = "serialize")]
 use serde::{Deserialize, Serialize};
 
 /// A [CompiledSchedule] is the output of the graph compiler.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "serialize", derive(Serialize, Deserialize))]
+#[derive(Clone, Debug)]
 pub struct CompiledSchedule {
     /// A list of nodes, delays, and summing points to
     /// evaluate in order to render audio, in topological order.
@@ -17,7 +20,8 @@ pub struct CompiledSchedule {
 }
 
 /// A [ScheduleEntry] is one element of the schedule to evalute.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "serialize", derive(Serialize, Deserialize))]
+#[derive(Clone, Debug)]
 pub enum ScheduleEntry {
     /// One of the input nodes, to process
     Node(ScheduledNode),
@@ -30,10 +34,11 @@ pub enum ScheduleEntry {
 
 /// A [ScheduledNode] is a [Node] that has been assigned buffers
 /// and a place in the schedule.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "serialize", derive(Serialize, Deserialize))]
+#[derive(Clone, Debug)]
 pub struct ScheduledNode {
     /// The unique ID of this node.
-    pub id: u64,
+    pub id: NodeID,
     /// The latency of this node. Kept for debugging and visualization.
     pub latency: f64,
     /// The assigned input buffers.
@@ -45,7 +50,8 @@ pub struct ScheduledNode {
 /// An [InsertedDelay] represents a required delay node to be inserted
 /// along some edge in order to compensate for different latencies along
 /// paths of the graph.
-#[derive(Copy, Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "serialize", derive(Serialize, Deserialize))]
+#[derive(Copy, Clone, Debug)]
 pub struct InsertedDelay {
     /// The edge that this delay corresponds to. Kept for debugging and visualization.
     pub edge: Edge,
@@ -60,7 +66,8 @@ pub struct InsertedDelay {
 /// An [InsertedSum] represents a point where multiple edges need to be merged
 /// into a single buffer, in order to support multiple inputs into the same
 /// port.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "serialize", derive(Serialize, Deserialize))]
+#[derive(Clone, Debug)]
 pub struct InsertedSum {
     /// The input buffers that will be summed
     pub input_buffers: Vec<BufferAssignment>,
@@ -70,7 +77,8 @@ pub struct InsertedSum {
 
 /// A [Buffer Assignment] represents a single buffer assigned to an input
 /// or output port.
-#[derive(Copy, Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "serialize", derive(Serialize, Deserialize))]
+#[derive(Copy, Clone, Debug)]
 pub struct BufferAssignment {
     /// The index of the buffer assigned
     pub buffer_index: usize,
@@ -80,9 +88,9 @@ pub struct BufferAssignment {
     /// passing it to a process
     pub should_clear: bool,
     /// The ID of the port this buffer is mapped to
-    pub port_id: u64,
+    pub port_id: PortID,
     /// The ID of the node this buffer is mapped to
-    pub node_id: u64,
+    pub node_id: NodeID,
     /// Buffers are reused, the "generation" represnts
     /// how many times this buffer has been used before
     /// this assignment. Kept for debugging and visualization.
