--- src/vec.rs
@@ -1,2 +1,0 @@
-use smallvec::SmallVec;
-pub type Vec<T> = SmallVec<[T; 16]>;
